Agents-Course-Learn/requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-community>=0.3.0  
langchain-core>=0.3.0

# LLM providers
langchain-groq>=0.2.0

# Environment utilities
python-dotenv>=1.0.0

# Tools (we'll add these later)
tavily-python>=0.5.0
langchain-huggingface>=0.3.0
langchain-chroma>=0.1.0
langgraph>=0.2.0